# swagger.yaml
openapi: 3.0.0
info:
  title: Sample API
  description: ""
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /user/all:
    get:
      tags:
        - user
      security:
        - bearerAuth: []
      summary: Returns all users in array
      description: Get all Users
      operationId: getAllUsers
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserAllResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /cart:
    post:
      tags:
        - Cart
      security:
        - bearerAuth: []
      summary: Return the cart object
      description: ""
      operationId: addCart
      requestBody:
        description: Product Details
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCart"
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /cart/view:
    get:
      tags:
        - Cart
      security:
        - bearerAuth: []
      summary: Return cart of user
      description: ""
      operationId: getCart
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetCartResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  /cart/update:
    patch:
      tags:
        - Cart
      security:
        - bearerAuth: []
      summary: Return Updated cart object
      description: ""
      operationId: updateCart

      requestBody:
        description: Product Details
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCart"
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateCartResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /wishlist/add:
    post:
      tags:
        - Wish-List
      security:
        - bearerAuth: []
      summary: Return Wishlist object
      description: ""
      operationId: addWishlist
      requestBody:
        description: product Details
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWishlist"
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WishList"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /product/create:
    post:
      tags:
        - Product
      security:
        - bearerAuth: []
      summary: Returns a Product object
      operationId: createProduct
      requestBody:
        description: Create Product
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProduct"
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /product/{id}:
    get:
      tags:
        - Product
      security:
        - bearerAuth: []
      summary: Return all Products in array
      description: ""
      operationId: getProductById
      parameters:
        - name: id
          in: path
          schema:
            type: number
          required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductByIdResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
  #orders
  /order:
    post:
      tags:
        - Order
      security:
        - bearerAuth: []
      summary: Return Order object
      description: ""
      operationId: createOrder
      requestBody:
        description: Order Details
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrder"
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /order/cancel/{id}:
    delete:
      tags:
        - Order
      security:
        - bearerAuth: []
      summary: Deletes order
      description: ""
      operationId: deleteProductById
      parameters:
        - name: id
          in: path
          schema:
            type: number
          required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelOrderResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  schemas:
...
    CreateCart:
      allOf:
        - type: object
          properties:
            product_id:
              type: number
              example: 1
            quantity:
              type: number
              example: "1"
    CreateWishlist:
      allOf:
        - type: object
          properties:
            product_id:
              type: number
              example: 1
    CreateProduct:
      allOf:
        - type: object
          properties:
            product_name:
              type: string
              example: product 1
            description:
              type: string
              example: Sample Description
            unit_price:
              type: number
              example: 100
            number_in_stock:
              type: number
              example: 500
    CreateOrder:
      allOf:
        - type: object
          properties:
            product:
              type: array
              items:
                $ref: "#/components/schemas/CreateCart"
    User:
      allOf:
        - type: object
          properties:
            user_id:
              type: number
              example: 1
            name:
              type: string
              example: jaiganeshan
            email:
              type: string
              example: gjai456@gmail.com
            password:
              type: string
              example: hU2FsdGVkX1/9EGiIHWWZoxRKk6GnpWHzxMbjg+GlZis=
            shipping_address:
              type: string
              example: null
            createdAt:
              type: string
              example: "2021-11-24T11:52:25.677Z"
            updatedAt:
              type: string
              example: "2021-11-24T11:52:25.677Z"
...
    UserAllResult:
      allOf:
        - type: array
          items:
            $ref: "#/components/schemas/User"
...
    Cart:
      allOf:
        - type: object
          properties:
            user_id:
              type: number
              example: 1
            product_id:
              type: number
              example: 1
            quantity:
              type: number
              example: 10
            createdAt:
              type: string
              example: "2021-11-24T11:52:25.677Z"
            updatedAt:
              type: string
              example: "2021-11-24T11:52:25.677Z"
    GetCartResult:
      allOf:
        - type: array
          items:
            type: object
            properties:
              product_id:
                type: number
                example: 1
              product_name:
                type: string
                example: product 1
              description:
                type: string
                example: description
              quantity:
                type: number
                example: 10
    UpdateCartResult:
      allOf:
        - type: object
          properties:
            message:
              type: string
              example: "Updated Successfully"
...
    WishList:
      allOf:
        - type: object
          properties:
            user_id:
              type: number
              example: 1
            product_id:
              type: number
              example: 1
            createdAt:
              type: string
              example: "2021-11-24T11:52:25.677Z"
            updatedAt:
              type: string
              example: "2021-11-24T11:52:25.677Z"
...
    Product:
      allOf:
        - type: object
          properties:
            product_id:
              type: number
              example: 1
            product_name:
              type: string
              example: product 1
            description:
              type: string
              example: Description about product
            unit_price:
              type: number
              example: 599
            number_in_stock:
              type: number
              example: 100
            seller_name:
              type: string
              example: Alex
            createdAt:
              type: string
              example: "2021-11-24T11:52:25.677Z"
            updatedAt:
              type: string
              example: "2021-11-24T11:52:25.677Z"
...
    ProductByIdResult:
      allOf:
        - type: object
        - $ref: "#/components/schemas/Product"
    #orderResult
    Order:
      allOf:
        - type: object
          properties:
            message:
              type: string
              example: "Order Placed Successfully"
    CancelOrderResult:
      allOf:
        - type: object
          properties:
            message:
              type: string
              example: "Order Cancelled"
...
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Error Message"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Unauthorized
    BadRequest:
      description: BadRequest
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Error Message"
    ServerError:
      description: BadRequest
      content:
        application/json:
          schema:
            type: string
            example: Something Failed!

  #security Schemes
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
#response Schemes